Added content fo derive command and to policies queries

+++ cmd/cmd_delete.go
@@ -1,10 +1,9 @@
 package cmd
 
 import (
+	db "github.com/Cloud-Pie/SPDT/storage"
 	"github.com/spf13/cobra"
 	"fmt"
-	"github.com/Cloud-Pie/SPDT/util"
-	db "github.com/Cloud-Pie/SPDT/storage"
 )
 
 // policiesCmd represents the delete policies command
@@ -12,19 +11,31 @@ var deleteCmd = &cobra.Command{
 	Use:   "delete",
 	Short: "Remove policy",
 	Long: "Remove a stored policy",
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Print("called delete")
-		id := ""
-		policyDAO := db.PolicyDAO{
-			Server:util.DEFAULT_DB_SERVER_POLICIES,
-			Database:util.DEFAULT_DB_POLICIES,
-		}
+	Run: delete,
+}
+
+var (
+	force bool
+	id	string
+)
+
+func delete(cmd *cobra.Command, args []string) {
+	if force {
+		policyDAO := db.GetPolicyDAO()
 		policyDAO.Connect()
 		err := policyDAO.DeleteById(id)
-
 		if err != nil {
 			fmt.Println("Error, policy could not be deleted")
-			fmt.Errorf(err.Error())
+			fmt.Println(err.Error())
+		} else {
+			fmt.Println("Policy deleted")
 		}
-	},
+	} else {
+		fmt.Println("Are you sure you want to delete this policy?, use the -f flag to force it")
+	}
+}
+
+func init() {
+	deleteCmd.Flags().StringVar(&id, "pId", "", "Policy ID")
+	deleteCmd.Flags().BoolVar(&force,"f", false, "Force the action")
 }
